requirements.txt
a2a-types==0.1.0
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.52.2
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
av==14.4.0
azure-core==1.34.0
azure-identity==1.23.0
blinker==1.9.0
cachetools==5.5.2
cartesia==2.0.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==45.0.3
dataclasses-json==0.6.7
deepgram-sdk==4.1.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.16
et_xmlfile==2.0.0
eval_type_backport==0.2.2
fast-agent-mcp==0.2.28
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.2
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-genai==1.19.0
googleapis-common-protos==1.70.0
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
instructor==1.8.3
iterators==0.2.0
Jinja2==3.1.6
jiter==0.8.2
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
livekit==1.0.8
livekit-agents==1.0.23
livekit-api==1.0.2
livekit-plugins-cartesia==1.0.23
livekit-plugins-deepgram==1.0.23
livekit-plugins-noise-cancellation==0.2.4
livekit-plugins-openai==1.0.23
livekit-plugins-silero==1.0.23
livekit-plugins-turn-detector==1.0.23
livekit-protocol==1.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.9.1
mdurl==0.1.2
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.41.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
onnxruntime==1.22.0
openai==1.84.0
openpyxl==3.1.5
opentelemetry-api==1.34.0
opentelemetry-distro==0.55b0
opentelemetry-exporter-otlp-proto-common==1.34.0
opentelemetry-exporter-otlp-proto-http==1.34.0
opentelemetry-instrumentation==0.55b0
opentelemetry-instrumentation-anthropic==0.40.7
opentelemetry-instrumentation-google-genai==0.2b0
opentelemetry-instrumentation-mcp==0.40.7
opentelemetry-instrumentation-openai==0.40.7
opentelemetry-proto==1.34.0
opentelemetry-sdk==1.34.0
opentelemetry-semantic-conventions==0.55b0
opentelemetry-semantic-conventions-ai==0.4.9
packaging==24.2
pandas==2.3.0
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.5
psutil==7.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
shellingham==1.5.4
silero==0.4.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
sse-starlette==2.3.6
starlette==0.46.2
streamlit==1.45.1
sympy==1.14.0
tenacity==9.1.2
tensorzero==2025.6.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
torchaudio==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.4
triton==3.3.1
typer==0.16.0
types-protobuf==4.25.0.20240417
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uuid_utils==0.11.0
uvicorn==0.34.3
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.22.0
